move test framework into its own directory
